Move calculations to useScaleCanvas hook
